+++ lib/src/lexemes/empty.dart
@@ -1,7 +1,14 @@
+import 'package:token_parser/src/debug.dart';
 import 'package:token_parser/src/lexeme.dart';
 import 'package:token_parser/src/token.dart';
 
 class EmptyLexeme extends Lexeme {
   @override
-  Token tokenize(String string, [ int start = 0 ]) => Token.emptyAt(this, string, start);
+  Token tokenize(String string, [ int start = 0 ]) {
+    DebugGrammar.debug(this, string, start);
+
+    return Token.emptyAt(this, string, start);
+  }
+
+  @override String get regexString => '';
 }