+++ lib/src/lexemes/pattern.dart
@@ -1,5 +1,5 @@
+import 'package:token_parser/src/debug.dart';
 import 'package:token_parser/src/error.dart';
-import 'package:token_parser/src/internal/extension.dart';
 import 'package:token_parser/src/lexeme.dart';
 import 'package:token_parser/src/lexemes/abstract/parent.dart';
 import 'package:token_parser/src/token.dart';
@@ -12,13 +12,15 @@ class PatternLexeme<PatternT extends Pattern> extends ParentLexeme {
 
   @override
   Token tokenize(String string, [ int start = 0 ]) {
+    DebugGrammar.debug(this, string, start);
+    
     final token = LexicalSyntaxError.enclose(this, () => pattern.matchAsPrefix(string, start));
     if (token == null) throw LexicalSyntaxError(this, string, start);
     return Token.match(this, token);
   }
 
   @override
-  String toString() => 
+  String get regexString => 
     pattern is RegExp
       ? (pattern as RegExp).pattern
     : pattern is String